Sorry for big messy commit.
I wrote the mlp agent.
I did a bunch of debugging, getting everything talking.

I ran into issue of music space rightly returning song nearest
to Agent recommendation which is not in the user taste for the
given user.

I think the best way to solve this is to hand the music space
the list of the songs the user has heard and have it return the
nearest one the user has heard. Maybe.