This is closer to working, but definitely broken.
I set the user to not change. The agent should at least be able
to learn one users song preferences.

The score gets lower while the loss does as well.
I think the theoretical justification for my update is not good.
I'm trying to use sampling from a normal distribution, but maybe
the direction from the recommended song to the actual song would
be better.

Need to come back and look at it later I think.